Add files via upload
Unfortunately I could not figure out how get the button Print the sorted words list after I clicked it. All of the examples given  are so simple. They are all system print line. I scoured the web and still could not fined out how to get it to display the list in the GUI. 
I messaged you about this early and I will continue to try and figure this out but as of right now I am stuck.